+++ docs/content/templates/category/index.mdx
@@ -2,7 +2,7 @@
 order: 2
 title: Category page
 
-group: Informational
+group: Website pages
 description: Our category page example.
 version: 1.0.0
 previewPath: /category
