+++ docs/content/templates/content/index.mdx
@@ -2,8 +2,8 @@
 order: 4
 title: Content page
 
-group: Informational
-description: The content template is used to deliver long-form informational content.
+group: Website pages
+description: The content template is used to deliver long-form Website pages content.
 version: 1.0.0
 previewPath: /category/subcategory/content
 figmaTemplateNodeId: 8410%3A83580
