+++ docs/content/templates/filtering-a-dataset/index.mdx
@@ -1,6 +1,6 @@
 ---
 title: Filtering a dataset
-group: Patterns
+group: Interactions
 description: Filters allows a user to refine the data they want to display in a large dataset by removing any items that do not match their criteria.
 version: 1.0.0
 figmaTemplateNodeId: 18586-56265
