+++ docs/content/templates/home/index.mdx
@@ -2,7 +2,7 @@
 order: 1
 title: Home page
 
-group: Informational
+group: Website pages
 description: Provide a strong first impression.
 version: 1.0.0
 previewPath: /
