+++ docs/content/templates/index.mdx
