+++ docs/content/templates/multi-page-form/index.mdx
@@ -1,7 +1,7 @@
 ---
 order: 6
 title: Multi-page form
-group: Forms
+group: Form pages
 description: Use this template to break up large forms into multiple smaller steps to decrease cognitive load.
 version: 1.0.0
 previewPath: /category/subcategory/multi-page-form
