+++ docs/content/templates/not-found/index.mdx
@@ -2,7 +2,7 @@
 order: 8
 title: Not found page
 
-group: Informational
+group: Website pages
 description: A standard 404 error page.
 version: 1.0.0
 previewPath: /404
