+++ docs/content/templates/selecting-multiple-options/index.mdx
@@ -1,6 +1,6 @@
 ---
 title: Selecting multiple options
-group: Patterns
+group: Interactions
 description: Multi-select is a commonly used design pattern that allows users to select multiple options from a list. This pattern is used in various contexts such as search filters, form fields, and more.
 version: 1.0.0
 ---
