+++ docs/content/templates/sign-in/index.mdx
@@ -2,7 +2,7 @@
 order: 7
 title: Sign-in form page
 
-group: Forms
+group: Form pages
 description: An example of a sign-in form.
 version: 1.0.0
 previewPath: /sign-in-form
