+++ docs/content/templates/single-page-form/index.mdx
@@ -2,7 +2,7 @@
 order: 5
 title: Single-page form
 
-group: Forms
+group: Form pages
 description: An example of a single-page multi-question form.
 version: 1.0.0
 previewPath: /category/subcategory/single-page-form
