+++ docs/content/templates/subcategory/index.mdx
@@ -2,7 +2,7 @@
 order: 3
 title: Subcategory page
 
-group: Informational
+group: Website pages
 description: Our subcategory page example.
 version: 1.0.0
 previewPath: /category/subcategory
