+++ docs/pages/patterns/index.tsx
@@ -1,75 +1,59 @@
-import { normalize } from 'path';
 import { Fragment } from 'react';
-import { Box, Stack } from '@ag.ds-next/react/box';
-import { Card, CardLink, CardInner } from '@ag.ds-next/react/card';
+import { Stack } from '@ag.ds-next/react/box';
 import { Columns } from '@ag.ds-next/react/columns';
-import { Text } from '@ag.ds-next/react/text';
-import { getMarkdownData } from '../../lib/mdxUtils';
-import { PATTERNS_PATH } from '../../lib/mdx/patterns';
+import { TextLink } from '@ag.ds-next/react/text-link';
+import { H2 } from '@ag.ds-next/react/heading';
+import { getTemplateGroupList, getTemplateList } from '../../lib/mdx/templates';
 import { DocumentTitle } from '../../components/DocumentTitle';
 import { CategoryPageTemplate } from '../../components/CategoryPageTemplate';
+import { TemplateCard } from '../../components/TemplateCard';
 
 type StaticProps = Awaited<ReturnType<typeof getStaticProps>>['props'];
 
-export default function PatternsPage({ title, description }: StaticProps) {
+export default function TemplatesPage({
+	groupList,
+	templateList,
+}: StaticProps) {
 	return (
-		<Fragment>
-			<DocumentTitle title={title} description={description} />
+		<>
+			<DocumentTitle title="Templates" />
 			<CategoryPageTemplate
-				title={title}
-				description={description}
-				editPath="/docs/content/patterns/index.mdx"
+				title="Pattern"
+				description="Patterns are reusable compositions of components that solve design problems and help ensure consistency across the service. They'll save your team time and resources and help get value to your users sooner."
+				editPath="/docs/pages/patterns/index.tsx"
 			>
 				<Stack gap={3}>
-					<Columns as="ul" gap={1.5} cols={{ xs: 1, sm: 2 }}>
-						<PatternCard
-							title="Patterns"
-							slug="patterns"
-							description="Patterns are reusable compositions of components that solve design problems and help ensure consistency across the service."
-						/>
-						<PatternCard
-							title="Templates"
-							slug="templates"
-							description="Templates can get your project up and running faster. They'll save your team time and resources and help get value to your users sooner."
-						/>
-					</Columns>
+					{[groupList[1], ...groupList.filter((_, idx) => idx !== 1)].map(
+						(group) => (
+							<Stack gap={1.5} key={group.slug}>
+								<H2 id={group.slug}>
+									<TextLink href={`/patterns/${group.slug}`}>
+										{group.title}
+									</TextLink>
+								</H2>
+								<Columns as="ul" gap={1.5} cols={{ xs: 1, sm: 2, lg: 3 }}>
+									{templateList
+										.filter((p) => p.group === group.slug)
+										.map((template) => (
+											<TemplateCard key={template.slug} {...template} />
+										))}
+								</Columns>
+							</Stack>
+						)
+					)}
 				</Stack>
 			</CategoryPageTemplate>
-		</Fragment>
-	);
-}
-
-function PatternCard({
-	title,
-	slug,
-	description,
-}: {
-	title: string;
-	slug: string;
-	description: string;
-}) {
-	return (
-		<Card as="li" clickable shadow>
-			<CardInner>
-				<Stack gap={1} flexGrow={1}>
-					<Box as="h3">
-						<CardLink href={`/patterns/${slug}`}>{title}</CardLink>
-					</Box>
-					<Text>{description}</Text>
-				</Stack>
-			</CardInner>
-		</Card>
+		</>
 	);
 }
 
 export async function getStaticProps() {
-	const { data } = await getMarkdownData(
-		normalize(`${PATTERNS_PATH}/index.mdx`)
-	);
+	const groupList = await getTemplateGroupList();
+	const templateList = await getTemplateList();
 	return {
 		props: {
-			title: data?.title as string,
-			description: data?.description as string,
+			groupList,
+			templateList,
 		},
 	};
 }
